# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for tracking workouts, exercises, and fitness routines",
        "title": "GetFit API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.yoursite.com/support",
            "email": "your.email@example.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "0.0.1"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/users/{userID}/routine/{routineID}/exercise/{exerciseID}": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Add an exercise with template sets to a workout routine",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "routine-exercises"
                ],
                "summary": "Add exercise to routine",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "userID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Routine ID",
                        "name": "routineID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Exercise ID",
                        "name": "exerciseID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Template sets information with version",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Exercise added to routine successfully",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Invalid request body or IDs",
                        "schema": {}
                    },
                    "500": {
                        "description": "Failed to add exercise to routine",
                        "schema": {}
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Remove an exercise from a workout routine",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "routine-exercises"
                ],
                "summary": "Remove exercise from routine",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "userID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Routine ID",
                        "name": "routineID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Exercise ID",
                        "name": "exerciseID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Expected version for optimistic concurrency",
                        "name": "version",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Exercise removed from routine successfully",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Invalid IDs or version",
                        "schema": {}
                    },
                    "500": {
                        "description": "Failed to remove exercise from routine",
                        "schema": {}
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Update template sets for an exercise in a routine",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "routine-exercises"
                ],
                "summary": "Update exercise in routine",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "userID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Routine ID",
                        "name": "routineID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Exercise ID",
                        "name": "exerciseID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Updated template sets with version",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Exercise template sets updated successfully",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Invalid request body or IDs",
                        "schema": {}
                    },
                    "500": {
                        "description": "Failed to update exercise in routine",
                        "schema": {}
                    }
                }
            }
        },
...
    },
...
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}